Revert "perf: use repository for webhooks list query & caching in /settings/developer/webhooks/... RSC (#21781)"
This reverts commit 0829929e983ab7438d6acbdd2e96d194ecf7f071.
