--- apps/web/app/(use-page-wrapper)/settings/(settings-layout)/developer/webhooks/[id]/actions.ts
@@ -1,7 +1,0 @@
-"use server";
-
-import { revalidateTag } from "next/cache";
-
-export async function revalidateWebhookById(id: string) {
-  revalidateTag(`viewer.webhook.get:${id}`);
-}
