--- apps/web/app/(use-page-wrapper)/settings/(settings-layout)/developer/webhooks/new/actions.ts
@@ -1,7 +1,0 @@
-"use server";
-
-import { revalidateTag } from "next/cache";
-
-export async function revalidateWebhookList() {
-  revalidateTag("viewer.webhook.list");
-}
