+++ apps/web/app/(use-page-wrapper)/settings/(settings-layout)/developer/webhooks/new/page.tsx
@@ -1,16 +1,10 @@
 import { createRouterCaller } from "app/_trpc/context";
 import { _generateMetadata } from "app/_utils";
-import { unstable_cache } from "next/cache";
-import { cookies, headers } from "next/headers";
-import { redirect } from "next/navigation";
 
-import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
 import { NewWebhookView } from "@calcom/features/webhooks/pages/webhook-new-view";
 import { APP_NAME } from "@calcom/lib/constants";
-import { WebhookRepository } from "@calcom/lib/server/repository/webhook";
 import { appsRouter } from "@calcom/trpc/server/routers/viewer/apps/_router";
-
-import { buildLegacyRequest } from "@lib/buildLegacyCtx";
+import { webhookRouter } from "@calcom/trpc/server/routers/viewer/webhook/_router";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
@@ -21,25 +15,15 @@ export const generateMetadata = async () =>
     "/settings/developer/webhooks/new"
   );
 
-const getCachedWebhooksList = unstable_cache(
-  async ({ userId }: { userId: number }) => {
-    return await WebhookRepository.findWebhooksByFilters({ userId });
-  },
-  undefined,
-  { revalidate: 3600, tags: ["viewer.webhook.list"] }
-);
-
 const Page = async () => {
-  const session = await getServerSession({ req: buildLegacyRequest(await headers(), await cookies()) });
-  if (!session?.user?.id) {
-    redirect("/auth/login");
-  }
-
-  const appsCaller = await createRouterCaller(appsRouter);
+  const [appsCaller, webhookCaller] = await Promise.all([
+    createRouterCaller(appsRouter),
+    createRouterCaller(webhookRouter),
+  ]);
 
   const [installedApps, webhooks] = await Promise.all([
     appsCaller.integrations({ variant: "other", onlyInstalled: true }),
-    getCachedWebhooksList({ userId: session.user.id }),
+    webhookCaller.list(),
   ]);
 
   return <NewWebhookView webhooks={webhooks} installedApps={installedApps} />;
